requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
psycopg2==2.9.10
pydantic==2.10.6
asyncpg==0.30.0
sqlalchemy==2.0.38
aiohttp==3.11.12
python-dotenv==1.0.1